+++ duckscript_sdk/src/sdk/std/collections/mod.rs
@@ -7,6 +7,8 @@ pub(crate) mod array_pop;
 mod array_push;
 mod is_array;
 mod map;
+mod map_get;
+mod map_put;
 mod range;
 mod read_properties;
 mod write_properties;
@@ -29,6 +31,8 @@ pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptEr
     commands.set(array_pop::create(&package))?;
     commands.set(is_array::create(&package))?;
     commands.set(map::create(&package))?;
+    commands.set(map_get::create(&package))?;
+    commands.set(map_put::create(&package))?;
     commands.set(range::create(&package))?;
     commands.set(read_properties::create(&package))?;
     commands.set(write_properties::create(&package))?;